pyproject.toml
[tool.poetry]
name = "musicdb"
version = "0.1.0"
description = "Your personal music database"
authors = ["tim <timkimcool@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
psycopg2-binary = "^2.9.6"
spotipy = "^2.23.0"
python-dotenv = "^1.0.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
django-allauth = "^0.54.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
